Merged PR 5117: Add JDK 19 to be built using gradle
Adds in new configs for JDK 19 for deb, rpm, suse-rpm. The configs include changes for JRE 19 and JDK 19. Most of these changes were based off of JDK 16 and JDK 17 changes. So they should be scrutinized a bit more to ensure nothing was left out or overlooked.

Bumps up the version allowed to be built to 19 in `linux/build.gradle`.

This doesnt include changes for JDK 18 (there might be an upstream commit which includes these though not 100% on that). If we desire to include JDK 18 we could add them before merging this commit.

Pipeline run: https://dev.azure.com/ms-juniper/Juniper/_build/results?buildId=33593&view=results

Related work items: #50266